Simplify block providers
