--- block/aws-beanstalk/provider.go
@@ -1,19 +1,0 @@
-package aws_beanstalk
-
-import (
-	"github.com/nullstone-io/deployment-sdk/aws/cloudwatch/metrics"
-	"github.com/nullstone-io/deployment-sdk/block"
-	"gopkg.in/nullstone-io/go-api-client.v0/types"
-)
-
-var ModuleContractName = types.ModuleContractName{
-	Category:    string(types.CategoryApp),
-	Subcategory: string(types.SubcategoryAppServer),
-	Provider:    "aws",
-	Platform:    "ec2",
-	Subplatform: "beanstalk",
-}
-
-var Provider = block.Provider{
-	NewMetricsGetter: metrics.NewGetter,
-}
